Fixed directory issues and organized files into folders. Also added Figure 1.
