--- Data/Analyzed_Micromotion_Data/8-16_Trial1_data.txt_analyzed.txt
@@ -1,1 +1,0 @@
-[-0.0008600626099927284, 210.0, 4.011621621621625]
