--- Data/Analyzed_Micromotion_Data/8-16_Trial2_data.txt_analyzed.txt
@@ -1,1 +1,0 @@
-[-0.0010147703178234344, 192.77777777777777, 4.200022176022174]
