+++ Data/Analyzed_Micromotion_Data/8-16_Trial2_data_analyzed.txt
@@ -1,0 +1,1 @@
+[-0.0010147703178234344, 192.77777777777777, 4.200022176022174]
