--- Data/Analyzed_Micromotion_Data/8-16_Trial3_data.txt_analyzed.txt
@@ -1,1 +1,0 @@
-[-0.0017991755371499238, 115.55555555555556, 4.3480649165824685]
