--- Data/Analyzed_Micromotion_Data/8-16_Trial4_data.txt_analyzed.txt
@@ -1,1 +1,0 @@
-[-0.0012258014455486476, 172.22222222222223, 4.385888340350015]
