+++ Data/Analyzed_Micromotion_Data/8-16_Trial4_data_analyzed.txt
@@ -1,0 +1,1 @@
+[-0.0012258014455486476, 172.22222222222223, 4.385888340350015]
