--- Data/Analyzed_Micromotion_Data/8-16_Trial5_data.txt_analyzed.txt
@@ -1,1 +1,0 @@
-[-0.0009092919238237538, 143.33333333333334, 3.303228571428571]
