--- Data/Analyzed_Micromotion_Data/8-16_Trial6_data.txt_analyzed.txt
@@ -1,1 +1,0 @@
-[-0.0010344684716006413, 177.22222222222223, 4.046378435635556]
