+++ Data/Analyzed_Micromotion_Data/8-16_Trial6_data_analyzed.txt
@@ -1,0 +1,1 @@
+[-0.0010344684716006413, 177.22222222222223, 4.046378435635556]
