+++ Data/Analyzed_Micromotion_Data/8-18_Trial11_data_analyzed.txt
@@ -1,0 +1,1 @@
+[-0.0010752576523061289, 189.44444444444446, 4.297965775401069]
