--- Data/Analyzed_Micromotion_Data/8-18_Trial12_data.txt_analyzed.txt
@@ -1,1 +1,0 @@
-[-0.0011267039569200048, 189.44444444444446, 4.413265873674603]
