+++ Data/Analyzed_Micromotion_Data/8-18_Trial12_data_analyzed.txt
@@ -1,0 +1,1 @@
+[-0.0011267039569200048, 189.44444444444446, 4.413265873674603]
