--- Data/Analyzed_Micromotion_Data/8-18_Trial13_data.txt_analyzed.txt
@@ -1,1 +1,0 @@
-[-0.001158468302308122, 202.77777777777777, 4.657133333087781]
