+++ Data/Analyzed_Micromotion_Data/8-18_Trial13_data_analyzed.txt
@@ -1,0 +1,1 @@
+[-0.001158468302308122, 202.77777777777777, 4.657133333087781]
