--- Data/Analyzed_Micromotion_Data/8-18_Trial15_data.txt_analyzed.txt
@@ -1,1 +1,0 @@
-[-0.0011169287803043674, 194.44444444444446, 4.456635736951525]
