--- Data/Analyzed_Micromotion_Data/8-18_Trial16_data.txt_analyzed.txt
@@ -1,1 +1,0 @@
-[-0.001021187116180137, 198.88888888888889, 4.29078058921961]
