+++ Data/Analyzed_Micromotion_Data/8-18_Trial16_data_analyzed.txt
@@ -1,0 +1,1 @@
+[-0.001021187116180137, 198.88888888888889, 4.29078058921961]
