+++ Data/Analyzed_Micromotion_Data/8-18_Trial17_data_analyzed.txt
@@ -1,0 +1,1 @@
+[-0.0010708209983238792, 207.77777777777777, 4.51763845485381]
