--- Data/Analyzed_Micromotion_Data/8-18_Trial18_data.txt_analyzed.txt
@@ -1,1 +1,0 @@
-[-0.001069234978408291, 204.44444444444446, 4.473007115928336]
