+++ Data/Analyzed_Micromotion_Data/8-18_Trial18_data_analyzed.txt
@@ -1,0 +1,1 @@
+[-0.001069234978408291, 204.44444444444446, 4.473007115928336]
