+++ Data/Analyzed_Micromotion_Data/8-18_Trial19_data_analyzed.txt
@@ -1,0 +1,1 @@
+[-0.0010483660287358696, 209.44444444444446, 4.484228144535082]
