--- Data/Analyzed_Micromotion_Data/8-18_Trial20_data.txt_analyzed.txt
@@ -1,1 +1,0 @@
-[-0.0010469826177468655, 200.55555555555554, 4.3725679091750935]
