+++ Data/Analyzed_Micromotion_Data/8-18_Trial20_data_analyzed.txt
@@ -1,0 +1,1 @@
+[-0.0010469826177468655, 200.55555555555554, 4.3725679091750935]
