+++ Data/Analyzed_Micromotion_Data/8-18_Trial7_data_analyzed.txt
@@ -1,0 +1,1 @@
+[-0.0011627319839532548, 202.12121212121212, 4.658252575305205]
