+++ Data/Analyzed_Micromotion_Data/8-18_Trial8_data_analyzed.txt
@@ -1,0 +1,1 @@
+[-0.001152185382844239, 187.77777777777777, 4.447077689025063]
