--- Data/Analyzed_Micromotion_Data/8-18_Trial9_data.txt_analyzed.txt
@@ -1,1 +1,0 @@
-[-0.0013398586459624805, 166.11111111111111, 4.518466462635493]
