--- Data/Analyzed_Micromotion_Data/8-8_Trial1_data.txt_analyzed.txt
@@ -1,1 +1,0 @@
-[-0.00124480407324099, 159.44444444444446, 4.234866515931885]
