+++ Data/Analyzed_Micromotion_Data/8-8_Trial1_data_analyzed.txt
@@ -1,0 +1,1 @@
+[-0.00124480407324099, 159.44444444444446, 4.234866515931885]
