+++ Data/Analyzed_Micromotion_Data/8-8_Trial2_data_analyzed.txt
@@ -1,0 +1,1 @@
+[-0.0010696349167425487, 182.22222222222223, 4.191254351372002]
